requirements.txt
Pillow>=9.1.1
tensorflow>=2.10.0
vaik-pascal-voc-rw-ex @ git+https://github.com/vaik-info/vaik-pascal-voc-rw-ex.git@644c9b53dcfa8a260a04fbfaf98ba25ee6a41d82